// __generated__/IssueListItem_issue.graphql.ts
export type IssueListItem_issue$data = {
  readonly created_at: any | null | undefined;
  readonly id: any;
  readonly issue_labelsCollection: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly labels: {
          readonly color: string | null | undefined;
          readonly id: any;
          readonly name: string | null | undefined;
        } | null | undefined;
      };
    }>;
  } | null | undefined;
  readonly nodeId: string;
  readonly priority: string | null | undefined;
  readonly status: string | null | undefined;
  readonly title: string | null | undefined;
  readonly users: {
    readonly avatar_url: string | null | undefined;
    readonly name: string | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "IssueListItem_issue";
};
export type IssueListItem_issue$key = {
  readonly " $data"?: IssueListItem_issue$data;
  readonly " $fragmentSpreads": FragmentRefs<"IssueListItem_issue">;
};